Move Handler into Boundary V1 namespace
