+++ cv-19-i-need-help-test/EndToEndTests/V1/GetHelpRequestsTest.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using Amazon.Lambda.APIGatewayEvents;
 using AutoFixture;
 using CV19INeedHelp;
+using CV19INeedHelp.Boundary.V1;
 using CV19INeedHelp.Boundary.V1.Responses;
 using CV19INeedHelp.Data.V1;
 using CV19INeedHelp.Helpers.V1;
