+++ cv-19-i-need-help/UseCases/V1/UpdateHelpRequestUseCase.cs
@@ -1,10 +1,5 @@
-using System;
-using System.Collections.Generic;
 using CV19INeedHelp.Gateways.V1;
 using CV19INeedHelp.Models.V1;
-using Amazon.Lambda.Core;
-using CV19INeedHelp.Boundary.V1.Responses;
-using Newtonsoft.Json;
 
 namespace CV19INeedHelp.UseCases.V1
 {
@@ -17,10 +12,10 @@ namespace CV19INeedHelp.UseCases.V1
             _iNeedHelpGateway = iNeedHelpGateway;
         }
 
-        public void UpdateHelpRequest(int id, ResidentSupportAnnexResponse data)
+        public void UpdateHelpRequest(int id, ResidentSupportAnnex data)
         {
             data.Id = id;
-            _iNeedHelpGateway.UpdateHelpRequest(data.ToEntity());
+            _iNeedHelpGateway.UpdateHelpRequest(data);
         }
 
         public void PatchHelpRequest(int id, ResidentSupportAnnexPatch data_items)
